+++ src/main/java/refinedstorage/block/BlockQuartzEnrichedIron.java
@@ -1,0 +1,7 @@
+package refinedstorage.block;
+
+public class BlockQuartzEnrichedIron extends BlockBase {
+    public BlockQuartzEnrichedIron() {
+        super("quartz_enriched_iron_block");
+    }
+}